Reconfiguration directory structure about input component (input, textarea)
